RavenDB-14881 Pull request rework. Changed data type from int to long (only for DetailedCollectionStatistics)

+++ src/Raven.Client/Documents/Operations/CollectionStatistics.cs
@@ -24,8 +24,8 @@ namespace Raven.Client.Documents.Operations
             Collections = new Dictionary<string, CollectionDetails>();
         }
 
-        public int CountOfDocuments { get; set; }
-        public int CountOfConflicts { get; set; }
+        public long CountOfDocuments { get; set; }
+        public long CountOfConflicts { get; set; }
 
         public Dictionary<string, CollectionDetails> Collections { get; set; }
     }
